Requirements Full Overhaul
Rewriteen all components of Requirements in proper syntax to increase clarity of the requirements.

The components of the functional requirements details its function based on the actors of the system, and also connects to the other component within the system.

Todo: In future changes need to put Non-Functional Requirements in its own section.

Added 7. AI component